Move Recipe class into the api.
