+++ src/main/java/mods/eln/api/recipe/Magnetizer.java
@@ -4,8 +4,6 @@ import java.util.ArrayList;
 
 import mods.eln.api.Misc;
 import net.minecraft.item.ItemStack;
-import mods.eln.Eln;
-import mods.eln.misc.Recipe;
 import mods.eln.misc.RecipesList;
 
 public class Magnetizer implements IELNMachine{
