+++ src/main/java/mods/eln/sim/ElectricalStackMachineProcess.java
@@ -1,6 +1,6 @@
 package mods.eln.sim;
 
-import mods.eln.misc.Recipe;
+import mods.eln.api.recipe.Recipe;
 import mods.eln.misc.RecipesList;
 import mods.eln.misc.Utils;
 import mods.eln.sim.mna.component.Resistor;
