+++ src/main/java/mods/eln/transparentnode/electricalmachine/ElectricalMachineDescriptor.java
@@ -3,6 +3,9 @@ package mods.eln.transparentnode.electricalmachine;
 import mods.eln.cable.CableRenderDescriptor;
 import mods.eln.gui.GuiLabel;
 import mods.eln.misc.*;
+import mods.eln.api.recipe.Recipe;
+import mods.eln.misc.RecipesList;
+import mods.eln.misc.Utils;
 import mods.eln.node.transparent.TransparentNodeDescriptor;
 import mods.eln.sim.ElectricalLoad;
 import mods.eln.sim.ElectricalStackMachineProcess;
